+++ packages/frontend/src/components/DashboardTiles/TileBase/index.tsx
@@ -10,8 +10,6 @@ import { Dashboard, DashboardTileTypes, isChartTile } from '@lightdash/common';
 import { Tooltip } from '@mantine/core';
 import { useHover, useToggle } from '@mantine/hooks';
 import React, { ReactNode, useState } from 'react';
-import { useParams } from 'react-router-dom';
-import { useChartSummaries } from '../../../hooks/useChartSummaries';
 import DeleteChartTileThatBelongsToDashboardModal from '../../common/modal/DeleteChartTileThatBelongsToDashboardModal';
 import ChartUpdateModal from '../TileForms/ChartUpdateModal';
 import TileUpdateModal from '../TileForms/TileUpdateModal';
@@ -53,7 +51,7 @@ const TileBase = <T extends Dashboard['tiles'][number]>({
     extraHeaderElement,
     titleHref,
 }: Props<T>) => {
-    const [isEditingChartTile, setIsEditingChartTile] = useState(false);
+    const [isEditingTileContent, setIsEditingTileContent] = useState(false);
     const [
         isDeletingChartThatBelongsToDashboard,
         setIsDeletingChartThatBelongsToDashboard,
@@ -63,8 +61,6 @@ const TileBase = <T extends Dashboard['tiles'][number]>({
     const { hovered: titleHovered, ref: titleRef } =
         useHover<HTMLAnchorElement>();
     const [isMenuOpen, toggleMenu] = useToggle([false, true]);
-    const { projectUuid } = useParams<{ projectUuid: string }>();
-    const { data: savedCharts } = useChartSummaries(projectUuid);
 
     const hideTitle =
         tile.type !== DashboardTileTypes.MARKDOWN
@@ -136,35 +132,9 @@ const TileBase = <T extends Dashboard['tiles'][number]>({
                                                                 icon="edit"
                                                                 text="Edit tile content"
                                                                 onClick={() =>
-                                                                    setIsEditingChartTile(
+                                                                    setIsEditingTileContent(
                                                                         true,
                                                                     )
-                                                                }
-                                                            />
-                                                        )}
-                                                        {tile.type !==
-                                                            DashboardTileTypes.MARKDOWN && (
-                                                            <MenuItem2
-                                                                icon={
-                                                                    hideTitle
-                                                                        ? 'eye-open'
-                                                                        : 'eye-off'
-                                                                }
-                                                                text={`${
-                                                                    hideTitle
-                                                                        ? 'Show'
-                                                                        : 'Hide'
-                                                                } title`}
-                                                                onClick={() =>
-                                                                    onEdit({
-                                                                        ...tile,
-                                                                        properties:
-                                                                            {
-                                                                                ...tile.properties,
-                                                                                hideTitle:
-                                                                                    !hideTitle,
-                                                                            },
-                                                                    })
                                                                 }
                                                             />
                                                         )}
@@ -213,49 +183,42 @@ const TileBase = <T extends Dashboard['tiles'][number]>({
                     <ChartContainer className="non-draggable sentry-block fs-block cohere-block">
                         {children}
                     </ChartContainer>
-                    {tile.type === DashboardTileTypes.SAVED_CHART ? (
-                        <ChartUpdateModal
-                            opened={isEditingChartTile}
-                            chartTitle={chartName ?? ''}
-                            onClose={() => setIsEditingChartTile(false)}
-                            onConfirm={(newTitle, newUuid) => {
-                                onEdit({
-                                    ...tile,
-                                    properties: {
-                                        ...tile.properties,
-                                        title:
-                                            newTitle.length > 0
-                                                ? newTitle
-                                                : savedCharts?.find(
-                                                      (chart) =>
-                                                          chart.uuid ===
-                                                          (newUuid.length > 0
-                                                              ? newUuid
-                                                              : tile.properties
-                                                                    .savedChartUuid),
-                                                  )?.name,
-                                        savedChartUuid:
-                                            newUuid.length > 0
-                                                ? newUuid
-                                                : tile.properties
-                                                      .savedChartUuid,
-                                    },
-                                });
-                                setIsEditingChartTile(false);
-                            }}
-                        />
-                    ) : (
-                        <TileUpdateModal
-                            className="non-draggable"
-                            isOpen={isEditingChartTile}
-                            tile={tile}
-                            onClose={() => setIsEditingChartTile(false)}
-                            onConfirm={(newTile) => {
-                                onEdit(newTile);
-                                setIsEditingChartTile(false);
-                            }}
-                        />
-                    )}
+                    {isEditingTileContent &&
+                        (tile.type === DashboardTileTypes.SAVED_CHART ? (
+                            <ChartUpdateModal
+                                opened={isEditingTileContent}
+                                tile={tile}
+                                onClose={() => setIsEditingTileContent(false)}
+                                onConfirm={(
+                                    newTitle,
+                                    newUuid,
+                                    shouldHideTitle,
+                                ) => {
+                                    onEdit({
+                                        ...tile,
+                                        properties: {
+                                            ...tile.properties,
+                                            title: newTitle,
+                                            savedChartUuid: newUuid,
+                                            hideTitle: shouldHideTitle,
+                                        },
+                                    });
+                                    setIsEditingTileContent(false);
+                                }}
+                                hideTitle={!!hideTitle}
+                            />
+                        ) : (
+                            <TileUpdateModal
+                                className="non-draggable"
+                                isOpen={isEditingTileContent}
+                                tile={tile}
+                                onClose={() => setIsEditingTileContent(false)}
+                                onConfirm={(newTile) => {
+                                    onEdit(newTile);
+                                    setIsEditingTileContent(false);
+                                }}
+                            />
+                        ))}
                     <DeleteChartTileThatBelongsToDashboardModal
                         className={'non-draggable'}
                         name={chartName ?? ''}