+++ packages/frontend/src/components/ProjectConnection/index.tsx
@@ -6,6 +6,7 @@ import {
     DbtProjectType,
     friendlyName,
     ProjectType,
+    SupportedDbtVersions,
     WarehouseTypes,
 } from '@lightdash/common';
 import { Anchor, Avatar, Card, Flex, Stack, Title } from '@mantine/core';
@@ -46,6 +47,7 @@ type ProjectConnectionForm = {
     dbt: DbtProjectConfig;
 
     warehouse?: CreateWarehouseCredentials;
+    dbtVersion: SupportedDbtVersions;
 };
 
 interface Props {
@@ -129,17 +131,6 @@ const ProjectForm: FC<Props> = ({
                         </Title>
                         <DocumentationHelpButton />
                     </LeftPanelTitle>
-
-                    <LeftPanelMessage>
-                        Your dbt project must be compatible with{' '}
-                        <Anchor
-                            href="https://docs.getdbt.com/docs/guides/migration-guide/upgrading-to-1-0-0"
-                            target="_blank"
-                            rel="noreferrer"
-                        >
-                            dbt version <b>1.4.* to 1.5.*</b>
-                        </Anchor>
-                    </LeftPanelMessage>
                 </div>
 
                 <div>
@@ -207,6 +198,7 @@ export const UpdateProjectConnection: FC<{
             warehouse: {
                 ...data?.warehouseConnection,
             },
+            dbtVersion: data?.dbtVersion,
         },
     });
     const { reset } = methods;
@@ -216,6 +208,7 @@ export const UpdateProjectConnection: FC<{
                 name: data.name,
                 dbt: data.dbtConnection,
                 warehouse: data.warehouseConnection,
+                dbtVersion: data.dbtVersion,
             });
         }
     }, [reset, data]);
@@ -225,6 +218,7 @@ export const UpdateProjectConnection: FC<{
         name,
         dbt: dbtConnection,
         warehouse: warehouseConnection,
+        dbtVersion,
     }: Required<ProjectConnectionForm>) => {
         if (user.data) {
             track({
@@ -234,6 +228,7 @@ export const UpdateProjectConnection: FC<{
                 name,
                 dbtConnection,
                 warehouseConnection,
+                dbtVersion: dbtVersion,
             });
         }
     };